terminado esboço da API e iniciado aulas de Angular

+++ Seção2 - Preparando a api do GetMestres/getMestres/getMestres.Api/src/controller/BaseController.ts
@@ -1,3 +1,4 @@
+//OBS.: as imports de Request, Response e NextFunction, são variáveis de ambiente
 import { Request } from "express"
 import { AppDataSource } from "../data-source"
 import { BaseNotification } from "../entity/BaseNotification"
@@ -8,16 +9,31 @@ import { Repository } from "typeorm"
 export abstract class BaseController<T> extends BaseNotification{
 
     private repository
+    private onlyRootController
+    private erroRoot: any = {
+        status: 401,
+        message: "Você não está autorizado a executar esta funcionalidade"
+    }
 
-    constructor(entity: any){
+    constructor(entity: any, onlyRoot: boolean = false){
         //Para evitar o erro no construtor, apenas declare o super() dentro do construtor
         //para chamar o construtor da super classe
         super()
         this.repository =  AppDataSource.getRepository<T>(entity)
+        this.onlyRootController = onlyRoot
+    }
+
+    //Este método irá verificar se o usuário é root ou não
+    //Caso seja, retornará true, caso não, retornará false e não deixará executar a função
+    private checkPermission(req: Request){
+        return this.onlyRootController && !req.IsRoot         
     }
 
     //Este controlador está acionando o GET
-    async all() {
+    async all(req: Request) {
+       if( this.checkPermission(req) ){
+        return this.erroRoot
+       }
         return this.repository.find({
             where: {
                 deleted: false
@@ -26,12 +42,15 @@ export abstract class BaseController<T> extends BaseNotification{
     }
 
     //Este controlador está acionando o GET com o paramêtro id
-    async one(request: Request) {
+    async one(req: Request) {
+        if( this.checkPermission(req) ){
+            return this.erroRoot
+        }
         //O paramêtro vem como id, mas a busca no banco de dados a chave primaria está como uid
-        const uid = request.params.id
+        const uid = req.params.id
         //para que a função findOne() funcione, será necessário incluir a claúsula where conforme abaixo        
         const user = await this.repository.findOne({
-            where: {uid}
+            where: {uid: uid}
         })
         //Caso não encontre o usuário do id informado, será exibido esta mensagem
         if (!user) {
@@ -44,19 +63,28 @@ export abstract class BaseController<T> extends BaseNotification{
     //Este é um método padrão para todos os controllers(pois seu model é any ou seja aceita qualquer classe que a chame)
     //E para ser usada, será necessário fazer uma sobrescrita do método
     //No controlador em questão e depois chamar este método que o super
-    async save(model: any) {        
+    async save(model: any, req: Request, ignorePermission: boolean = false) {
+
+        if( !ignorePermission ){
+            if( this.checkPermission(req) ){
+                return this.erroRoot
+            }
+        }
+                
         if( model.uid ){
-            delete model['deleted']            
-            delete model['createAt']
-            delete model['updateAt']
+            delete model.deleted            
+            delete model.createAt
+            delete model.updateAt
 
             let _modelInDB = await this.repository.findOne({
                 where: {uid: model.uid}
+                //update tableName set a, b, c, d where uid = model.uid
             })
             if( _modelInDB ){
                 Object.assign(_modelInDB, model)
             }
         }
+        
 
         if( this.valid() ){
             return this.repository.save(model)
@@ -70,14 +98,19 @@ export abstract class BaseController<T> extends BaseNotification{
     }
     
     //Este método irá apagar todos os dados da entidade em questão
-    async apagaTudo(){
+    async apagaTudo(req: Request){
+        if( this.checkPermission(req) ){
+            return this.erroRoot
+        }
         await this.repository.delete({active: 1})
         return "Todos os dados foram apagados com sucesso"
     }
     
     async remove(req: Request){
-        var texto = ""
-        
+        if( this.checkPermission(req) ){
+            return this.erroRoot
+        }
+
         const uid = req.params.id
         //Primeiro recupera do banco de dados a informação
         var info = await this.repository.findOne({
@@ -98,8 +131,11 @@ export abstract class BaseController<T> extends BaseNotification{
 
     //Este controlador está acionando o DELETE com o paramêtro id
     //Este método irá apenas desabilitar o usuário do banco, como se estiver apagado
-    async disable(request: Request) {
-        let uid = request.params.id
+    async disable(req: Request) {
+        if( this.checkPermission(req) ){
+            return this.erroRoot
+        }
+        let uid = req.params.id
         let model = await this.repository.findOne({
             where: {
                 uid: uid
